TravelBlog-48: Updated laravel to 6.x.

+++ database/migrations/2019_09_16_110900_create_posts_table.php
@@ -15,9 +15,9 @@ class CreatePostsTable extends Migration
     public function up()
     {
         Schema::create('posts', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
-            $table->integer('category_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned();
+            $table->bigInteger('category_id')->unsigned();
             $table->string('status')->nullable();
             $table->string('isPopular')->nullable();
             $table->longText('header')->nullable();
