+++ database/migrations/2019_09_26_122810_create_followers_table.php
@@ -15,10 +15,10 @@ class CreateFollowersTable extends Migration
     public function up()
     {
         Schema::create('followers', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('follower_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('follower_id')->unsigned();
             $table->foreign('follower_id')->references('id')->on('users')->onDelete('cascade');
-            $table->integer('following_id')->unsigned();
+            $table->bigInteger('following_id')->unsigned();
             $table->foreign('following_id')->references('id')->on('users')->onDelete('cascade');
             $table->string('created_at');
         });
